# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "gestion_usuarios",
        "description": "Proyecto API REST hecho en Laravel para la gestión de usuarios de una base de datos MySQL",
        "contact": {
            "email": "juanmanuel.pernia@ara-tech.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:80/gestion_usuarios/public/"
        }
    ],
    "paths": {
...
        "/api/rols": {
            "get": {
                "tags": [
                    "Rols"
                ],
                "summary": "List all rols",
                "description": "List all rols",
                "operationId": "Listrols",
                "responses": {
                    "200": {
                        "description": "List of all successful roles"
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    },
                    "404": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rols"
                ],
                "summary": "Store a new rol in the database",
                "description": "Store a new rol",
                "operationId": "7e2a4f8207a168bb91aab147fb527ee7",
                "requestBody": {
                    "description": "rol to store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the rol",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol successfully created"
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    },
                    "422": {
                        "description": "Error: Unprocessable Content"
                    },
                    "500": {
                        "description": "Error: Internal Server Error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/rols/{id}": {
            "get": {
                "tags": [
                    "Rols"
                ],
                "summary": "Show a specific rol",
                "description": "Show a specific rol",
                "operationId": "713e6cc28efc82d2760134c2ad410370",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the rol to show",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol found"
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    },
                    "404": {
                        "description": "Error: Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rols"
                ],
                "summary": "Update a specific rol",
                "description": "Update a specific rol",
                "operationId": "18c650c41c659b8ea6282004e8be9305",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the rol to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Rol to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The name of the rol",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol successfully modified"
                    },
                    "404": {
                        "description": "Error: Not Found"
                    },
                    "422": {
                        "description": "Error: Unprocessable Content"
                    },
                    "500": {
                        "description": "Error: Internal Server Error"
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rols"
                ],
                "summary": "Delete a specific rol",
                "description": "Delete a specific rol",
                "operationId": "d6eafcb80350ca803baa733349bef536",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the rol to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol successfully eliminated "
                    },
                    "401": {
                        "description": "Error: Unauthorized"
                    },
                    "404": {
                        "description": "Error: Not found"
                    },
                    "500": {
                        "description": "Error: Internal Server Error"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "API Key Auth",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}